requirements.txt
# TC Dashboard — Dependencies

# Core Framework
dash>=2.17.0
dash-mantine-components>=0.14.0
dash-iconify>=0.1.2
dash-leaflet>=1.0.15

# Data Processing
pandas>=2.2.0
numpy>=1.26.0
plotly>=5.18.0

# Caching (optional — app runs without Redis)
redis>=5.0.0

# Utilities
python-dotenv>=1.0.0
pytz>=2024.1

# Production
gunicorn>=22.0.0
